Added test case for combined account and standalone solr server
